bpf: Add an egress BPF program to the native device
The program will perform reverse NAT when direct server return is needed
for packets originating from local processes.
