+++ bpf/Makefile
@@ -7,7 +7,7 @@ CLANG_FLAGS :=  ${FLAGS} -target bpf
 # let clang complain too early.
 CLANG_FLAGS += -Wall -Werror -Wno-address-of-packed-member -Wno-unknown-warning-option
 
-BPF = bpf_lxc.o bpf_netdev.o bpf_overlay.o
+BPF = bpf_lxc.o bpf_netdev.o bpf_overlay.o bpf_netdev_tx.o
 SCRIPTS = init.sh join_ep.sh run_probes.sh
 LIB := $(shell find ./lib -name '*.h')
