+++ bpf/lib/maps.h
@@ -52,22 +52,6 @@ struct bpf_elf_map __section_maps cilium_reserved_policy = {
 	.max_elem	= RESERVED_POLICY_SIZE,
 };
 
-struct bpf_elf_map __section_maps cilium_proxy4 = {
-	.type		= BPF_MAP_TYPE_HASH,
-	.size_key	= sizeof(struct proxy4_tbl_key),
-	.size_value	= sizeof(struct proxy4_tbl_value),
-	.pinning	= PIN_GLOBAL_NS,
-	.max_elem	= 8192,
-};
-
-struct bpf_elf_map __section_maps cilium_proxy6= {
-	.type		= BPF_MAP_TYPE_HASH,
-	.size_key	= sizeof(struct proxy6_tbl_key),
-	.size_value	= sizeof(struct proxy6_tbl_value),
-	.pinning	= PIN_GLOBAL_NS,
-	.max_elem	= 8192,
-};
-
 /* Private per EP map for internal tail calls */
 struct bpf_elf_map __section_maps CALLS_MAP = {
 	.type		= BPF_MAP_TYPE_PROG_ARRAY,
