+++ daemon/bpf.sha
@@ -1,2 +1,2 @@
-GO_BINDATA_SHA1SUM=d93f75aca108215012cf1a126e31ae4068dce3ab
+GO_BINDATA_SHA1SUM=419446f19cb49ab280753f2aca1e22c68d921640
 GO_VERSION_USED=go1.8.3
