+++ daemon/daemon.go
@@ -546,6 +546,9 @@ func (d *Daemon) init() error {
 	ipv4GW := nodeaddress.GetInternalIPv4()
 	fmt.Fprintf(fw, "#define IPV4_GATEWAY %#x\n", byteorder.HostSliceToNetwork(ipv4GW, reflect.Uint32).(uint32))
 
+	publicIP4 := nodeaddress.GetExternalIPv4()
+	fmt.Fprintf(fw, "#define IPV4_PUBLIC_IP %#x\n", byteorder.HostSliceToNetwork(publicIP4, reflect.Uint32).(uint32))
+
 	if !d.conf.IPv4Disabled {
 		fmt.Fprintf(fw, "#define IPV4_LOOPBACK %#x\n", byteorder.HostSliceToNetwork(d.loopbackIPv4, reflect.Uint32).(uint32))
 	}
